Clean up utils.js
